updated the measurement results for different execution types
