+++ src/nipcbatt/pcbatt_library/dcpower/common/__init__.py
@@ -1,0 +1,1 @@
+"""Provides common helper functions for DC power operations."""
